// src/app/core/interface/policy-inputs.interface.ts
export interface PolicyInput {
  quote_id: string;
  policy_id: string;
  account_loc_id: number;
  rating_period: string;
  policy_status: string;
  policy_type: string;
  effective_date: string;
  expiration_date: string;
  pro_rate_factor: string;
  entity_name: string;
  address: string;
  city: string;
  state: string;
  zip_code: string;
  county: string;
  original_year_built: number;
  construction_description: string;
  sprinklered: string;
  type_of_wiring: string;
  num_pools: number;
  num_playgrounds: number;
  num_courts_gyms_other: number;
  pct_occupied: number;
  pct_subsidized: string;
  pct_student: string;
  pct_assisted_living: string;
  property_type: string;
  occupancy_details: string;
  num_buildings: number;
  num_stories: number;
  num_hab_units: number;
  commercial_sqft: number;
  habitational_sqft: number;
  parking_sqft: number;
  total_sqft: number;
  fein_tin: string;
  bvd_id: string;
  location_description: string;
  location_address: string;
  occ_agg_limit: string;
  deductible: number;
  primary_operation: string;
  gl_class_code_primary: string;
  gl_primary_exposure_basis: string;
  gl_primary_exposure: number;
  gl_secondary_class_code_rating: string;
  gl_secondary_operation: string;
  gl_secondary_class_code: string;
  gl_secondary_exposure_basis: string;
  gl_secondary_exposure: number;
  gl_tertiary_class_code_rating: string;
  gl_tertiary_operation: string;
  gl_tertiary_class_code: string;
  gl_tertiary_exposure_basis: string;
  gl_tertiary_exposure: number;
  primary_parking_type: string;
  primary_parking_class_code: string;
  primary_parking_exposure_basis: string;
  primary_parking_exposure: number;
  secondary_parking_type: string;
  secondary_parking_class_code: string;
  secondary_parking_exposure_basis: string;
  secondary_parking_exposure: number;
  tertiary_parking_type: string;
  tertiary_parking_class_code: string;
  tertiary_parking_exposure_basis: string;
  tertiary_parking_exposure: number;
  waiver_of_subrogation: string;
  primary_and_noncontributory: string;
  blanket_additional_insured: string;
  hired_and_nonowned_auto: string;
  assault_and_battery_sub_limit: string;
  habitation_sub_limit: string;
  fire_alarm: string;
  common_areas_adequate_lighting: string;
  driveways_parking_common_areas_good_condition: string;
  emergency_exit_signs: string;
  formal_safety_disaster_plan: string;
  formal_employee_training_program: string;
  formal_maintenance_safety_work_order_tracking: string;
  safety_equipment_under_10_years: string;
  pets_pet_restrictions: string;
  security_cameras_coverage: string;
  smoke_detectors_all_units: string;
  walkways_sidewalks_hallways_good_condition: string;
}

export interface PolicyInputsResponse {
  policy_inputs_list: PolicyInput[];
}